src/main/java/cn/isnap/leetcode/algorithm/level2/day07/SubtreeOfAnotherTree.java
package cn.isnap.leetcode.algorithm.level2.day07;

import cn.isnap.leetcode.TreeNode;

public class SubtreeOfAnotherTree {
    public boolean isSubtree(TreeNode root, TreeNode subRoot) {
        if (root == null) return false;
        if (same(root, subRoot)) return true;
        return isSubtree(root.left, subRoot) || isSubtree(root.right, subRoot);
    }

    private boolean same(TreeNode root, TreeNode subRoot) {
        if (root == null && subRoot == null) return true;
        if (root == null || subRoot == null) return false;
        if (root.val != subRoot.val) return false;

        return same(root.left, subRoot.left) && same(root.right, subRoot.right);
    }
}
